Support log messages in psitest, too
