+++ libraries/psibase/native/include/psibase/LogSocket.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <memory>
+
+namespace psibase
+{
+   struct Socket;
+   std::shared_ptr<Socket> makeLogSocket();
+}  // namespace psibase
